Dropbox extractor (v2) update - continue
This commit contains an major overhaul in Dropbox extractor. This new version gets rid of a need for an separate OAuth2 external authorization and use a third-party certified Dropbox module. By getting rid of these extra authorization hassle, the whole code was simplified and stood practically the same as before.